chore(test): move test file next to source
