+++ index.js
@@ -1,6 +1,6 @@
-exports.Metrics = require('./Metrics');
-exports.StringReporter = require('./reporters/StringReporter');
-exports.ConsoleReporter = require('./reporters/ConsoleReporter');
-exports.GraphiteReporter = require('./reporters/GraphiteReporter');
-exports.DataDogReporter = require('./reporters/DataDogReporter');
-exports.InMemoryReporter = require('./reporters/InMemoryReporter');
+exports.Metrics = require('./metrics');
+exports.StringReporter = require('./reporters/string-reporter');
+exports.ConsoleReporter = require('./reporters/console-reporter');
+exports.GraphiteReporter = require('./reporters/graphite-reporter');
+exports.DataDogReporter = require('./reporters/datadog-reporter');
+exports.InMemoryReporter = require('./reporters/in-memory-reporter');
