+++ package.json
@@ -13,7 +13,7 @@
     "url": "https://github.com/ysa23/metrics-js"
   },
   "scripts": {
-    "test": "mocha test",
+    "test": "mocha --recursive *.test.js **/*.test.js",
     "lint": "eslint index.js .",
     "example:graphite" : "node ./examples/graphite.js"
   },
