+++ reporters/string-reporter.js
