+++ reporters/string-reporter.test.js
